# src/docs/swagger.yaml
definitions:
...
  model.EventSeatCategories:
    properties:
      eventID:
        type: string
      price:
        type: integer
      seatCategoryID:
        type: string
    type: object
  model.EventSeats:
    properties:
      blockedUntil:
        type: string
      booked:
        type: boolean
      eventID:
        type: string
      id:
        type: string
      seatID:
        type: string
      userID:
        type: string
    type: object
  model.Events:
    properties:
      cinemaHallID:
        type: string
      description:
        type: string
      end:
        type: string
      eventType:
        type: string
      id:
        type: string
      is3d:
        type: boolean
      start:
        type: string
      title:
        type: string
    type: object
  model.Genres:
    properties:
      genreName:
        type: string
      id:
        type: string
    type: object
  model.Movies:
    properties:
      bannerPicURL:
        type: string
      coverPicURL:
        type: string
      description:
        type: string
      fsk:
        type: integer
      id:
        type: string
      rating:
        type: number
      releaseDate:
        type: string
      timeInMin:
        type: integer
      title:
        type: string
      trailerURL:
        type: string
    type: object
...
  model.SeatCategories:
    properties:
      categoryName:
        type: string
      id:
        type: string
    type: object
  model.Seats:
    properties:
      cinemaHallID:
        type: string
      columnNr:
        type: integer
      id:
        type: string
      rowNr:
        type: integer
      seatCategoryID:
        type: string
      type:
        type: string
    type: object
...
  models.CreateEvtDTO:
    properties:
      cinemaHallID:
        type: string
      description:
        type: string
      end:
        type: string
      eventSeatCategories:
        items:
          $ref: '#/definitions/model.EventSeatCategories'
        type: array
      eventType:
        type: string
      id:
        type: string
      is3d:
        type: boolean
      movies:
        items:
          type: string
        type: array
      start:
        type: string
      title:
        type: string
    type: object
...
  models.GenreWithMovies:
    properties:
      genreName:
        type: string
      id:
        type: string
      movies:
        items:
          properties:
            bannerPicURL:
              type: string
            coverPicURL:
              type: string
            description:
              type: string
            fsk:
              type: integer
            id:
              type: string
            rating:
              type: number
            releaseDate:
              type: string
            timeInMin:
              type: integer
            title:
              type: string
            trailerURL:
              type: string
          type: object
        type: array
    type: object
...
  models.GetEventSeatsDTO:
    properties:
      eventSeat:
        $ref: '#/definitions/model.EventSeats'
      eventSeatCategory:
        $ref: '#/definitions/model.EventSeatCategories'
      seat:
        $ref: '#/definitions/model.Seats'
      seatCategory:
        $ref: '#/definitions/model.SeatCategories'
    type: object
  models.GetEventSeatsResponse:
    properties:
      blockedUntil:
        type: string
      currentUserSeats:
        items:
          $ref: '#/definitions/models.GetSeatsForSeatSelectorDTO'
        type: array
      seat_rows:
        items:
          items:
            $ref: '#/definitions/models.GetSeatsForSeatSelectorDTO'
          type: array
        type: array
    type: object
...
  models.GetSeatsForSeatSelectorDTO:
    properties:
      available:
        type: boolean
      bookedByOther:
        type: boolean
      category:
        type: string
      columnNr:
        type: integer
      id:
        type: string
      price:
        type: integer
      rowNr:
        type: integer
      type:
        type: string
    type: object
  models.GetSelectedSeatsResponse:
    properties:
      selectedSeats:
        items:
          $ref: '#/definitions/models.GetEventSeatsDTO'
        type: array
    type: object
  models.GetSpecialEventsDTO:
    properties:
      cinemaHallID:
        type: string
      description:
        type: string
      end:
        type: string
      eventType:
        type: string
      id:
        type: string
      is3d:
        type: boolean
      movies:
        items:
          $ref: '#/definitions/model.Movies'
        type: array
      start:
        type: string
      title:
        type: string
    type: object
...
  models.KTSErrorMessage:
    properties:
      errorMessage:
        type: string
    type: object
...
  models.PatchEventSeatResponse:
    properties:
      blockedUntil:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /events:
    post:
      consumes:
      - application/json
      description: Create event
      parameters:
      - description: Event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.CreateEvtDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create event
      tags:
      - Events
  /events/{eventId}:
    get:
      consumes:
      - application/json
      description: Get event by id
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetSpecialEventsDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get event by id
      tags:
      - Events
...
  /events/{eventId}/seats:
    get:
      consumes:
      - application/json
      description: Get event seats
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetEventSeatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get event seats
      tags:
      - EventSeats
  /events/{eventId}/seats/{seatId}/block:
    patch:
      consumes:
      - application/json
      description: Block event seat
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PatchEventSeatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Block event seat
      tags:
      - EventSeats
  /events/{eventId}/seats/{seatId}/unblock:
    patch:
      consumes:
      - application/json
      description: Unblock event seat
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PatchEventSeatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Unblock event seat
      tags:
      - EventSeats
  /events/{eventId}/user-seats:
    get:
      consumes:
      - application/json
      description: Get selected seats
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetSelectedSeatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get selected seats
      tags:
      - EventSeats
  /events/special:
    get:
      consumes:
      - application/json
      description: Get special events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetSpecialEventsDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get special events
      tags:
      - Events
  /genres:
    get:
      consumes:
      - application/json
      description: Get genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Genres'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get genres
      tags:
      - Genres
    post:
      consumes:
      - application/json
      description: Create genre
      parameters:
      - description: Genres model
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/model.Genres'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Genres'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Create genre
      tags:
      - Genres
    put:
      consumes:
      - application/json
      description: Update genre
      parameters:
      - description: Genres model
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/model.Genres'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Genres'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Update genre
      tags:
      - Genres
  /genres/{id}:
    delete:
      consumes:
      - application/json
      description: Delete genre
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Delete genre
      tags:
      - Genres
  /genres/{name}:
    get:
      consumes:
      - application/json
      description: Get genre by name
      parameters:
      - description: Genre name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Genres'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get genre by name
      tags:
      - Genres
  /genres/movies:
    get:
      consumes:
      - application/json
      description: Get genres with movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GenreWithMovies'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get genres with movies
      tags:
      - Genres
...
  /movies/{id}/events:
    get:
      consumes:
      - application/json
      description: Get events for movie
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Events'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.KTSErrorMessage'
      summary: Get events for movie
      tags:
      - Events
...
swagger: "2.0"